+++ habit_flow/lib/features/habit/domain/usecases/add_habit_usecase.dart
@@ -11,14 +11,7 @@ class AddHabitUseCase {
   final HabitRepository _repository;
 
   Future<Either<Failure, Habit>> execute(String title) async {
-    // TODO: 1. Validate that the title is not empty. If empty, return Left(ValidationFailure('Title cannot be empty')).
     if (title.isEmpty) return left(const ValidationFailure('Title cannot be empty'));
-
-    // TODO: 2. Create a specific Habit object.
-    // - id: use Uuid().v4()
-    // - title: the validated title
-    // - createdAt: DateTime.now()
-    // - isCompleted: false
 
     final habit = Habit(
       id: const Uuid().v4(),